Fri, Dec 23, 2022, 6:52 PM +00:00

+++ credit/credit.c
@@ -9,10 +9,10 @@ int main(void)
 //Master: 51-55
 //Visa:4
 int check_digit=0;
-int i = 15.0;
+double i = 15.0;
     do
     {
-        if (card_no % pow(10,i)==1)
+        if (card_no % pow(10.0,i)==1)
             {
                 check_digit = 1
             }
